I think I fixed the ln bug

+++ support/interface.py
@@ -56,7 +56,11 @@ def doMove(name, season, episode, src):
     ext = src[src.find('.')+1:]
     srcFile = '/home/pi/drive/torrents/Anime-bot/complete/{}'.format(src)
     destFile = '/home/pi/drive/Media/Anime/{0}/Season {1}/{0} - s{1}e{2}.{3}'.format(name, rjust(season), rjust(episode), ext)
-    shutil.move(srcFile, destFile)
+    if os.path.isFile(destFile):
+        return #Hopefully this keeps me from invalidating 800Mb of traffic limit
+    else:
+        shutil.move(srcFile, destFile)
+        os.sym(destFile, srcFile)
 
 def rjust(num):
     return '0'*(2-len(str(num)))+str(num)
